invisible-watermark/requirements.txt
torch>=1.0.0
torchvision>=0.2.0
numpy>=1.16.0
Pillow>=6.0.0
tensorboardX>=1.8
tensorboard>=1.14
opencv-python>=4.0.0
scipy>=1.2.0
matplotlib>=3.0.0
tqdm>=4.30.0
